+++ Runtime/Transitioners/Modules/ColorRgbRendererModule.cs
@@ -53,11 +53,13 @@ namespace HiddenAchievement.CrossguardUi.Modules
         }
 
         /// <inheritdoc />
-        public void Transition(RectTransform component, IStyleModuleRule rule, float duration)
+        public void Transition(RectTransform component, IStyleModuleRule rule, float duration, Ease easing)
         {
             if (rule is not ColorRgbRendererModuleRule colorRule) return;
             CanvasRenderer renderer = _componentCache[component];
-            _motionHandle = LMotion.Create(renderer.GetColor(), colorRule.Color, duration).BindToColorNoAlpha(renderer);
+            _motionHandle = LMotion.Create(renderer.GetColor(), colorRule.Color, duration)
+                .WithEase(easing)
+                .BindToColorNoAlpha(renderer);
         }
 
         /// <inheritdoc />